+++ CHANGELOG.md
@@ -9,12 +9,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- AI Guard: detector overrides.
-- AI Guard: topic detector.
-- AI Guard: `ignore_recipe` in detector overrides.
 - Management: new API client.
 - Redact: config APIs.
 - Secure Audit Log: config APIs.
+
+## 5.1.0 - 2025-04-25
+
+### Added
+
+- AuthZ: `expiresAt` to tuples.
+- AuthN: groups.
+
+## 5.0.0 - 2025-04-21
+
+### Added
+
+- AI Guard: detector overrides.
+- AI Guard: topic detector.
+- AI Guard: `ignore_recipe` in detector overrides.
+- `baseUrlTemplate` has been added to `Config` to allow for greater control over
+  the complete API URL. This option may be a full URL with the optional
+  `{SERVICE_NAME}` placeholder, which will be replaced by the slug of the
+  respective service name. This supersedes `environment` and `insecure`.
 
 ### Changed